better handling of enclosure, improved fetching and deleting of content images

+++ engine/ocitems.cpp
@@ -254,7 +254,13 @@ void OcItems::itemsRequestedUpdateDb(QVariantMap requestItemsResult, QString typ
         updateEventFeed(newEventItems);
 
     if (imgHandling == 2 && !newItems.isEmpty())
+    {
+        // wait one second before fetching images to give the database models some time to load
+        QEventLoop loop;
+        QTimer::singleShot(1000, &loop, SLOT(quit()));
+        loop.exec();
         fetchImages(newItems);
+    }
 }
 
 
@@ -575,7 +581,14 @@ void OcItems::itemsUpdatedUpdateDb(QVariantMap updateItemsResult, QString type, 
         updateEventFeed(newEventItems);
 
     if (imgHandling == 2 && !newItems.isEmpty())
+    {
+        // wait one second before fetching images to give the database models some time to load
+        QEventLoop loop;
+        QTimer::singleShot(1000, &loop, SLOT(quit()));
+        loop.exec();
         fetchImages(newItems);
+    }
+
 }
 
 
@@ -1475,36 +1488,7 @@ void OcItems::deleteCachedImages(const QList<int> &idsToDelte)
 {
     QSqlQuery query;
 
-//    for (int d = 0; d < idsToDelte.size(); ++d)
-//    {
-
-//        QStringList nameFilter;
-//        QString fileFilter("_*");
-//        fileFilter.prepend(QString::number(idsToDelte.at(d)));
-//        nameFilter << fileFilter;
-//        QString storagePath(QDir::homePath());
-//        storagePath.append(IMAGE_CACHE);
-//        QDir directory(storagePath);
-//        QStringList imgs = directory.entryList(nameFilter);
-
-//    #ifdef QT_DEBUG
-//        qDebug() << "Images to delete: " << imgs;
-//    #endif
-
-//        for (int i = 0; i < imgs.size(); ++i)
-//        {
-//            directory.remove(imgs.at(i));
-//        }
-//    }
-
     QStringList imagesToDelete;
-
-//    for (int d = 0; d < idsToDelte.size(); ++d)
-//    {
-//        query.exec(QString("SELECT path FROM images WHERE parentId = %1").arg(idsToDelte.at(d)));
-//        while(query.next())
-//            imagesToDelete << query.value(0).toString();
-//    }
 
     QSqlDatabase::database().transaction();
     for (int i = 0; i < idsToDelte.size(); ++i)
@@ -1517,9 +1501,12 @@ void OcItems::deleteCachedImages(const QList<int> &idsToDelte)
     }
     QSqlDatabase::database().commit();
 
-    for (int f = 0; f < imagesToDelete.size(); ++f)
+    if (!imagesToDelete.isEmpty())
     {
-        QFile::remove(imagesToDelete.at(f));
+        for (int f = 0; f < imagesToDelete.size(); ++f)
+        {
+            QFile::remove(imagesToDelete.at(f));
+        }
     }
 }
